making a function to render a menu

+++ src/drawing.h
@@ -10,7 +10,7 @@ extern "C" {
 #endif
 
 void color_pixel(union rgb color, int x, int y);
-void drawRectangle(union rgb color, int x, int y, int width, int height);
+void drawRectangle(union rgb color, int x, int y, int16_t width, uint16_t height);
 void drawSquare(int x, int y);
 void drawShape(int (*shapeMatrix)[4], int y, int x);
 int (*rotateLeft(int (*matrix)[4]))[4];
